refactor: 폴더 구조 변경
- repository 인터페이스, 구현 클래스 도입을 위해 pointhistory 폴더와 userpoint 폴더 생성
- pointhistory 폴더와 userpoint 폴더에는 각 테이블에 대응하는 리포지토리 인터페이스, 구현 클래스 추가될 예정
